requirements.txt
dnspython==2.6.1
Requests==2.31.0
shodan==1.31.0
tldextract==5.1.1